Physics based animation in C++
The following is a physics based animation that I did in C++ using Lagrangian mechanics. The animation itself consists of two components, a stick and a sphere. The sphere has a hole going down the center which allows the stick to fall through it. The stick itself has two smaller balls in each end which serve as blockers so the stick doesn't completely fall through the sphere. The main sphere is placed on a slope and begins to roll down. The stick is placed through the center of the sphere where the hole is. As the sphere rolls down the hill, the stick continuously falls through which each half rotation. Eventually, the sphere reaches a velocity such that the stick appears to stay in place due to the high velocities.